+++ documentation-website/src/client/Packages/ReactPkg.js
@@ -43,7 +43,7 @@ router.respond((response) => {
     <CuriBase
       response={response}
       router={router}
-      render={response => {
+      render={({ response }) => {
         return response.body ? <response.body /> : null;
       }}
     />
@@ -59,9 +59,9 @@ router.respond((response) => {
           <Subsection tag="h4" title="render" id="CuriBase-render">
             <p>
               A render function. This will be called whenever the{" "}
-              <Cmp>CuriBase</Cmp> renders. The function will be passed the
-              current response object and the router object it was passed as a
-              prop. The function must return a React element.
+              <Cmp>CuriBase</Cmp> renders. The function will be passed an object
+              with three properties: <IJS>response</IJS>, <IJS>navigation</IJS>,
+              and <IJS>router</IJS>. The function must return a React element.
             </p>
           </Subsection>